♻️ refactor other pr input into dedicated component
